+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
-## v0.1.0
+## v0.2.0 - 20170111
+
+- Add tests & coverage
+- Add es6 distribution
+- Add docs
+- Add ButtonLink component
+
+## v0.1.0 - 20170108
 
 - Add Link component